feat: Adds Dev Container configuration
Includes the reference image, the relevant extensions, and the commands
to run (installing the dependencies)
And any relevant ports to forward.

It also adds the Welcome message to appear afterwards.

Next commit will be adding the Workflow to Pre-Build the image to make
it faster to start the development environment.